+++ ed128d.hpp
@@ -11,6 +11,6 @@
 class Double128 : public EuclideanDistance
 {
 public:
-    double calculate(const double *x,const double *y, size_t n) override;
+    double calculate(const double *x,const double *y, int n) override;
 };
 #endif //ED128D_HPP
