+++ ednaive.cpp
@@ -5,7 +5,7 @@
 #include "ednaive.hpp"
 
 
-double Baseline::calculate(const double *x, const double *y, size_t n) {
+double Baseline::calculate(const double *x, const double *y, int n) {
     double result = 0;
     for (; n > 0; --n) {
         const double num = x[n] - y[n];
