+++ ednaive.hpp
@@ -10,7 +10,7 @@
 class Baseline : public EuclideanDistance
 {
 public:
-    double calculate(const double *x,const double *y, size_t n) override;
+    double calculate(const double *x,const double *y, int n) override;
 };
 
 #endif //EDNAIVE_HPP
